Move GlobalSetup to ProcessLib.
It is more related to the NumericsConfig than to the DOF table.
